refactor: move model to its own package
